Support of large files (4+ Gb) on windows. ftell and fseek use long, which is only 32 bits on windows. Need to use ftelli64 and fseeki64, especially with SEEK_SET. This will be needed when we use container files.

+++ inc/oapv.h
@@ -738,8 +738,8 @@ struct oapv_multi_mip_decode {
 typedef struct oapvd_istream oapvd_istream_t;
 struct oapvd_istream {
     void *data;
-    long (*tell)(oapvd_istream_t *bitr);
-    int (*seek)(oapvd_istream_t *bitr, long offset, int origin);
+    long long (*tell)(oapvd_istream_t *bitr);
+    int (*seek)(oapvd_istream_t *bitr, long long offset, int origin);
     size_t (*read)(oapvd_istream_t *bitr, void *buffer, size_t size, size_t count);
 };
